add path from capture to vision api
