--- controllers/vision_controller.js
@@ -1,11 +1,0 @@
-// const db = require("../models");
-const express = require("express");
-const router = express.Router();
-
-// GET route to get a specific user's beer and review
-router.get("/api/vision", (req, res) => {
-  console.log(req);
-});
-
-
-module.exports = router;
